Improve ptrace support.
Only i386 Linux syscalls are supported and, even then, not all of them.
Stuff was moved around and we are now faking some file descriptors to
refer to their respective driver. Logging exists but it kind of sucks.

Note: amd64 programs will not run unless they use the i386 syscall
interface. It's possible to disassemble the code at %rip to know if we
used syscall or int $0x80 but there's a race condition: another thread
might overwrite this memory while we're handling it.